requirements.txt
opencv-python
opencv-contrib-python
numpy
cupy-cuda12x
pyrealsense2
